requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Image processing and OCR
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
pypdf==3.17.4

# File format support
python-multipart==0.0.6
python-magic==0.4.27

# HTTP client and async support
httpx==0.25.2
aiofiles==23.2.1
redis[hiredis]==5.0.1

# Data processing
pydantic==2.5.0
numpy==1.25.2
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0

# Environment and configuration
python-dotenv==1.0.0

# Process management
supervisor==4.2.5

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
respx==0.20.2